Add tests for new PluginsManager methods
